requirements_music_search.txt
# music_search 패키지용 (기존 requirements.txt 와 별도)
# 설치: pip install -r requirements_music_search.txt

numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
librosa>=0.10.0
soundfile>=0.12
numba>=0.57
matplotlib>=3.7

# FAISS: Windows 에서 faiss-cpu 권장 (없으면 numpy 브루트포스로 동작)
faiss-cpu>=1.7.4; platform_system != "Darwin" or platform_machine != "arm64"
# Apple Silicon 등은: conda install faiss-cpu 또는 pip 시도

# 자연어 질의 → 다국어 임베딩 (선택, 없으면 TF-IDF만)
sentence-transformers>=2.2.0
torch>=2.0.0
